Changed how rules are handled when undoing actions
Old state:
Each rule key is saved in the undo buffer. when user undo the rule is deleted. broblem is if 2 action performed on same element the first undo that deleted the rule also deleted the first action
New state:
Each rule key  with the modified detail is saved in the undo buffer. when user undo that detail is unset. only is the entire rule has no value (fall details are false/null) it will be deleted. broblem is that undoing first of 2 coloring actions also delete the rule.