ruleActions: Rename setCssSelector(), reorder the action's function call
No point in performing actions on an element that is to be removed. Best but the removeElement() first so consecutive calls on same element will not find it and wil no nothing.
The isolate actions may turn the selector invalid (parent element removed) so its best to color it before isolating it. the colorization wil be copied.